feat(moodle-websvc): Fetching from Moodle works! :smile:
Problem was posting form-data to the service.
Workaround was sending as URL params.
Trouble is that those GET params are visible on any requests, thus exposing the security key.
For now, for just a single user, I'm not worried, but I do want to be sure to fix this in the future.